# src/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"Task": {
				"properties": {
					"id": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"assignedAssistant": {
						"type": "string"
					},
					"status": {
						"type": "string",
						"enum": [
							"pending",
							"in_progress",
							"completed",
							"failed"
						]
					},
					"inputData": {
						"type": "string",
						"nullable": true
					},
					"outputData": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"description",
					"assignedAssistant",
					"status",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Tag": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"MemoryWithTags": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"summary": {
						"type": "string",
						"nullable": true
					},
					"type": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"data": {
						"type": "string",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"tags": {
						"items": {
							"$ref": "#/components/schemas/Tag"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"id",
					"name",
					"summary",
					"type",
					"description",
					"data",
					"createdAt",
					"updatedAt",
					"tags"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MemoryFocusRule": {
				"properties": {
					"id": {
						"type": "string"
					},
					"assistantId": {
						"type": "string"
					},
					"maxResults": {
						"type": "number",
						"format": "double"
					},
					"relationshipTypes": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"priorityTags": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"assistantId",
					"maxResults",
					"relationshipTypes",
					"priorityTags",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Feedback": {
				"properties": {
					"id": {
						"type": "string"
					},
					"targetId": {
						"type": "string"
					},
					"targetType": {
						"type": "string",
						"enum": [
							"assistant",
							"memory",
							"task"
						]
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"comments": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"targetId",
					"targetType",
					"rating",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Pick_Feedback.Exclude_keyofFeedback.id__": {
				"properties": {
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"targetId": {
						"type": "string"
					},
					"targetType": {
						"type": "string",
						"enum": [
							"assistant",
							"memory",
							"task"
						]
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"comments": {
						"type": "string"
					}
				},
				"required": [
					"createdAt",
					"updatedAt",
					"targetId",
					"targetType",
					"rating"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"Omit_Feedback.id_": {
				"$ref": "#/components/schemas/Pick_Feedback.Exclude_keyofFeedback.id__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"Partial_Omit_Feedback.id-or-createdAt-or-updatedAt__": {
				"properties": {
					"targetId": {
						"type": "string"
					},
					"targetType": {
						"type": "string",
						"enum": [
							"assistant",
							"memory",
							"task"
						]
					},
					"rating": {
						"type": "number",
						"format": "double"
					},
					"comments": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
			"Assistant": {
				"properties": {
					"id": {
						"type": "string"
					},
					"gptAssistantId": {
						"type": "string",
						"nullable": true
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"chat",
							"assistant"
						]
					},
					"model": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"type",
					"model",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"FeedbackSummary": {
				"properties": {
					"avgRating": {
						"type": "number",
						"format": "double"
					},
					"totalFeedback": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"avgRating",
					"totalFeedback"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AssistantWithDetails": {
				"properties": {
					"id": {
						"type": "string"
					},
					"gptAssistantId": {
						"type": "string",
						"nullable": true
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"chat",
							"assistant"
						]
					},
					"model": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					},
					"focusedMemories": {
						"items": {
							"$ref": "#/components/schemas/MemoryWithTags"
						},
						"type": "array"
					},
					"memoryFocusRule": {
						"$ref": "#/components/schemas/MemoryFocusRule"
					},
					"assistantTags": {
						"items": {
							"$ref": "#/components/schemas/Tag"
						},
						"type": "array"
					},
					"feedbackSummary": {
						"$ref": "#/components/schemas/FeedbackSummary"
					}
				},
				"required": [
					"id",
					"name",
					"description",
					"type",
					"model",
					"createdAt",
					"updatedAt",
					"focusedMemories",
					"feedbackSummary"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Partial_Assistant_": {
				"properties": {
					"id": {
						"type": "string"
					},
					"gptAssistantId": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"assistant",
							"chat"
						]
					},
					"model": {
						"type": "string"
					},
					"createdAt": {
						"type": "string"
					},
					"updatedAt": {
						"type": "string"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "gpt-assistants-api",
		"version": "0.2.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/task/assistant/{assistantId}": {
			"get": {
				"operationId": "GetTasksByAssistant",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Task"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Task"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "assistantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/memory-owned/assistant/{assistantId}": {
			"get": {
				"operationId": "GetMemoriesByAssistantId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MemoryWithTags"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"MemoryOwned"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "assistantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateOwnedMemories",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"MemoryOwned"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "assistantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"memoryIds": {
										"items": {
											"type": "string"
										},
										"type": "array"
									}
								},
								"required": [
									"memoryIds"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
		"/memory-focused/assistant/{assistantId}": {
			"get": {
				"operationId": "GetFocusedMemoriesByAssistantId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MemoryWithTags"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"MemoryFocused"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "assistantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/memory-focus-rule/assistant/{assistantId}": {
			"get": {
				"operationId": "GetMemoryFocusRules",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MemoryFocusRule"
								}
							}
						}
					}
				},
				"tags": [
					"MemoryFocusRule"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "assistantId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/memory-extra/by-tags": {
			"get": {
				"operationId": "GetMemoriesByTags",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/MemoryWithTags"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"MemoryExtra"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "tags",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/memory-extra/{memoryId}/tags": {
			"put": {
				"operationId": "UpdateMemoryTags",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"MemoryExtra"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "memoryId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"newTags": {
										"items": {
											"type": "string"
										},
										"type": "array"
									}
								},
								"required": [
									"newTags"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/feedback/{id}": {
			"get": {
				"operationId": "GetFeedbackById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Feedback"
								}
							}
						}
					}
				},
				"tags": [
					"Feedback"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateFeedback",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Feedback"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Omit_Feedback.id-or-createdAt-or-updatedAt__"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteFeedback",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Feedback"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/feedback/target/{targetType}/{targetId}": {
			"get": {
				"operationId": "GetFeedbackByTarget",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Feedback"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Feedback"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "targetType",
						"required": true,
						"schema": {
							"type": "string",
							"enum": [
								"assistant",
								"memory",
								"task"
							]
						}
					},
					{
						"in": "path",
						"name": "targetId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/feedback": {
			"post": {
				"operationId": "AddFeedback",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "string"
										}
									},
									"required": [
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Feedback"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Omit_Feedback.id_"
							}
						}
					}
				}
			}
		},
...
		"/assistant": {
			"get": {
				"operationId": "GetAllAssistants",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Assistant"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Assistant"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateAssistant",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "string"
										}
									},
									"required": [
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Assistant"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"instructions": {
										"type": "string"
									},
									"model": {
										"type": "string"
									},
									"type": {
										"type": "string",
										"enum": [
											"chat",
											"assistant"
										]
									},
									"description": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"instructions",
									"model",
									"type",
									"description",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/assistant/{id}": {
			"get": {
				"operationId": "GetAssistantById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Assistant"
								}
							}
						}
					}
				},
				"tags": [
					"Assistant"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateAssistant",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Assistant"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_Assistant_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteAssistant",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Assistant"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/assistant/{id}/details": {
			"get": {
				"operationId": "GetAssistantWithDetailsById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AssistantWithDetails"
								}
							}
						}
					}
				},
				"tags": [
					"Assistant"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/assistant/simple": {
			"post": {
				"operationId": "CreateAssistantSimple",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"id": {
											"type": "string"
										}
									},
									"required": [
										"id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Assistant"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"instructions": {
										"type": "string"
									},
									"name": {
										"type": "string"
									}
								},
								"required": [
									"instructions",
									"name"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/"
		}
	]
}